first pass at getting wit.ai to help us with times!
