+++ app/jobs/send_twilio_messages_job.rb
@@ -48,7 +48,7 @@ class SendTwilioMessagesJob < Struct.new(:messages, :phone_numbers, :smsCampaign
             @outgoing.body = message
             @outgoing.from = ENV['TWILIO_SURVEY_NUMBER'].gsub('+1', '').delete('-')
             @outgoing.wufoo_formid = smsCampaign
-            # @incoming.direction = "incoming-twiml"
+            @outgoint.direction = 'outgoing-survey'
             @outgoing.save
 
             phone_number = '+1' + phone_number.strip.gsub('+1', '').delete('-')
@@ -68,7 +68,7 @@ class SendTwilioMessagesJob < Struct.new(:messages, :phone_numbers, :smsCampaign
             Rails.logger.warn("[Twilio][SendTwilioMessagesJob] had a problem. Full error: #{@outgoing.error_message}")
           end
         end
-        sleep(1)
+        sleep(1) # for twilio rate limiting, I presume.
       end
     end
   end
