+++ app/sms/reservation_cancel_sms.rb
@@ -9,20 +9,15 @@ class ReservationCancelSms < ApplicationSms
     @reservation = reservation
   end
 
-  def send
-    client.messages.create(
-      from: application_number,
-      to:   to.phone_number,
-      body: "The #{duration} minute interview for #{selected_time}, with #{reservation.user.name} has been cancelled")
+  def body
+    "The #{duration} minute interview for #{selected_time}, with #{reservation.user.name} has been cancelled"
   end
 
-  private
+  def selected_time
+    reservation.time_slot.start_datetime_human
+  end
 
-    def selected_time
-      reservation.time_slot.start_datetime_human
-    end
-
-    def duration
-      reservation.duration / 60
-    end
+  def duration
+    reservation.duration / 60
+  end
 end
