+++ config/initializers/delayed_job_config.rb
@@ -1,6 +1,6 @@
 # config/initializers/delayed_job_config.rb
 Delayed::Worker.destroy_failed_jobs = false
-Delayed::Worker.sleep_delay = 60
+Delayed::Worker.sleep_delay = 1 # sms needs to be speedy
 Delayed::Worker.max_attempts = 3
 Delayed::Worker.max_run_time = 5.minutes
 Delayed::Worker.read_ahead = 10
