+++ config/initializers/redis.rb
@@ -1,0 +1,1 @@
+Redis.current = Redis.new(:host => 'localhost', :port => 6379)
