+++ lib/extensions/with_user.rb
